NB-132 : refactor - UserFollow 관련 Controller, Service 리팩토링
- Controller에서는 parameter와 관련된 처리에 집중
- Service에서 다른 객체와의 통신 책임 구현

+++ src/main/java/com/soyeon/nubim/domain/userfollow/UserFollowService.java
@@ -2,12 +2,15 @@ package com.soyeon.nubim.domain.userfollow;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.soyeon.nubim.domain.user.User;
 import com.soyeon.nubim.domain.user.UserMapper;
 import com.soyeon.nubim.domain.user.UserService;
 import com.soyeon.nubim.domain.user.dto.UserSimpleResponseDto;
+import com.soyeon.nubim.domain.userfollow.dto.FollowUserResponseDto;
 
 import lombok.RequiredArgsConstructor;
 
@@ -18,7 +21,19 @@ public class UserFollowService {
 	private final UserService userService;
 	private final UserMapper userMapper;
 
-	public void createFollow(User follower, User followee) {
+	public FollowUserResponseDto createFollow(Long userId) {
+		userService.validateUserExists(userId);
+		User follower = userService.getCurrentUser();
+		User followee = userService.findUserByIdOrThrow(userId);
+
+		if (follower.equals(followee)) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Cannot follow yourself");
+		}
+
+		if (this.isFollowing(follower, followee)) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "You are already following");
+		}
+
 		UserFollow userFollow = UserFollow.builder()
 			.follower(follower)
 			.followee(followee)
@@ -26,6 +41,12 @@ public class UserFollowService {
 
 		userFollowRepository.save(userFollow);
 		userService.addFollowerAndFolloweeByUserFollow(userFollow);
+
+		return FollowUserResponseDto.builder()
+			.followerId(userFollow.getFollower().getUserId())
+			.followeeId(userFollow.getFollowee().getUserId())
+			.message("Successfully followed")
+			.build();
 	}
 
 	public boolean isFollowing(User follower, User followee) {
@@ -36,21 +57,39 @@ public class UserFollowService {
 		}
 	}
 
-	public void deleteUserFollow(User follower, User followee) {
+	public FollowUserResponseDto deleteUserFollow(Long userId) {
+		userService.validateUserExists(userId);
+
+		User follower = userService.getCurrentUser();
+		User followee = userService.findUserByIdOrThrow(userId);
+
+		if (!isFollowing(follower, followee)) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "You are not following");
+		}
 		UserFollow userFollowToDelete = userFollowRepository.findByFollowerAndFollowee(follower, followee)
 			.orElseThrow();
 		userFollowRepository.delete(userFollowToDelete);
 		userService.deleteFollowerAndFolloweeByUserFollow(userFollowToDelete);
+
+		return FollowUserResponseDto.builder()
+			.followerId(follower.getUserId())
+			.followeeId(followee.getUserId())
+			.message("Successfully unfollowed")
+			.build();
 	}
 
-	public Page<UserSimpleResponseDto> getFollowers(User followee, Pageable pageable) {
+	public Page<UserSimpleResponseDto> getFollowers(Pageable pageable) {
+		User followee = userService.getCurrentUser();
+
 		Page<UserFollow> filteredUserFollows = userFollowRepository.findByFollowee(followee, pageable);
 
 		return filteredUserFollows.map(
 			userFollow -> userMapper.toUserSimpleResponseDto(userFollow.getFollower()));
 	}
 
-	public Page<UserSimpleResponseDto> getFollowees(User follower, Pageable pageable) {
+	public Page<UserSimpleResponseDto> getFollowees(Pageable pageable) {
+		User follower = userService.getCurrentUser();
+
 		Page<UserFollow> filteredUserFollows = userFollowRepository.findByFollower(follower, pageable);
 
 		return filteredUserFollows.map(
